SearchEngine.py
This is a customized search engine build with the help of python, and the tools which are used are exa API to access the LLM and streamlit to build an user interface. To run the programyou need to run the following command in powershell in order to access the interface python -m streamlit run "your address of the python file".